Renamed MGDL_PLATFORM_WINDOWS to MGDL_PLATFORM_MSYS2. Removed using libpng and libvorbis and ogg on pc platform. Those are done with stb now

+++ include/mgdl/mgdl-sound.h
@@ -30,7 +30,7 @@
 	};
 	typedef enum SoundSampleFormat SoundSampleFormat;
 
-	short Sound_FormatToChannels(SoundSampleFormat);
+	int Sound_FormatToChannels(SoundSampleFormat);
 //! Sound handling struct
 /*!
  *	\details Class object for loading and playing back sound effects with pitch control and stereo panning.
